% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvntest.h.R
\name{mvntest}
\alias{mvntest}
\title{Multivariate Normality Test}
\usage{
mvntest(
  data,
  vars,
  group = NULL,
  mvnTest = "hz",
  univariateTest = "AD",
  showDescriptives = TRUE,
  outlierMethod = "none",
  outlierAlpha = 0.05,
  transform = "none",
  powerFamily = "none",
  scale = FALSE,
  impute = "none",
  bootstrap = FALSE,
  nBoot = 1000,
  showQQPlot = FALSE,
  showUniPlots = FALSE,
  showBoxPlots = FALSE,
  showHistograms = FALSE
)
}
\arguments{
\item{data}{.}

\item{vars}{Numeric variables to test for multivariate normality.}

\item{group}{Optional grouping variable for stratified analysis.}

\item{mvnTest}{The multivariate normality test to use.}

\item{univariateTest}{The univariate normality test to apply to each
variable.}

\item{showDescriptives}{Show descriptive statistics table.}

\item{outlierMethod}{Method for detecting multivariate outliers using
robust Mahalanobis distance.}

\item{outlierAlpha}{Significance level for outlier detection.}

\item{transform}{Marginal transformation to apply to each variable.}

\item{powerFamily}{Power transformation family to apply before analysis.}

\item{scale}{Standardize data (zero mean, unit variance) before analysis.}

\item{impute}{Method for handling missing data.}

\item{bootstrap}{Use bootstrap resampling for p-value estimation. The
Energy test always uses parametric bootstrap, regardless of this setting.}

\item{nBoot}{Number of bootstrap replicates. This value is also used by the
Energy test.}

\item{showQQPlot}{Show chi-square Q-Q plot for multivariate normality
assessment.}

\item{showUniPlots}{Show univariate Q-Q plots for each variable.}

\item{showBoxPlots}{Show box plots for each variable.}

\item{showHistograms}{Show histograms with normal curve overlay.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$mvnTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$uniTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$descTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$outlierTable} \tab \tab \tab \tab \tab a table \cr
  \code{results$qqPlot} \tab \tab \tab \tab \tab an image \cr
  \code{results$uniPlots} \tab \tab \tab \tab \tab an image \cr
  \code{results$boxPlots} \tab \tab \tab \tab \tab an image \cr
  \code{results$histPlots} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$mvnTable$asDF}

\code{as.data.frame(results$mvnTable)}
}
\description{
Multivariate Normality Test
}
